+++ 02-express-tutorial/logger.js
@@ -1,0 +1,9 @@
+const logger = (req, res, next) =>{
+    const method = req.method
+    const url = req.url
+    const time = new Date().getFullYear()
+    console.log(method, url, time)
+    next()
+}
+
+module.exports = logger
